Crate collectives_westend_integration_tests

Modules

collectives_fellowship

The Westend Technical Fellowship.

opaque

v3

Version 3 of the Cross-Consensus Message format data structures.

westend_governance

New governance configurations for the Kusama runtime.

Macros

assert_expected_events

bx

test_parachain_is_trusted_teleporter

Structs

AlwaysV3

WrapVersion implementation which attempts to always convert the XCM to version 3 before wrapping it.

AlwaysV4

WrapVersion implementation which attempts to always convert the XCM to version 3 before wrapping it.

AlwaysV5

WrapVersion implementation which attempts to always convert the XCM to version 3 before wrapping it.

Ancestor

A unit struct which can be converted into a Location of the inner parents value.

AncestorThen

A unit struct which can be converted into a Location of the inner parents value and the inner interior.

Asset

Either an amount of a single fungible asset, or a single well-identified non-fungible asset.

AssetHubWestendReceiver

AssetHubWestendSender

AssetHubWestendXcmConfig

AssetId

Location to identify an asset.

Assets

A Vec of Assets.

CollectivesWestendReceiver

CollectivesWestendSender

CollectivesWestendXcmConfig

HintNumVariants

Location

A relative path between state-bearing consensus systems.

PalletInfo

Parent

A unit struct which can be converted into a Location of parents value 1.

ParentThen

A tuple struct which can be converted into a Location of parents value 1 with the inner interior.

QueryResponseInfo

Information regarding the composition of a query response.

Weight

WestendReceiver

WestendSender

WestendXcmConfig

Xcm

XcmContext

Contextual data pertaining to a specific list of XCM instructions.

Enums

AssetFilter

Asset collection, defined either by a number of Assets or a single wildcard.

AssetInstance

A general identifier for an instance of a non-fungible asset class.

BodyId

An identifier of a pluralistic body.

BodyPart

A part of a pluralistic body.

Fungibility

Classification of whether an asset is fungible or not, along with a mandatory amount or instance.

Hint

Instruction

Cross-Consensus Message: A message from one consensus system to another.

Junction

A single item in a path to describe the relative location of a consensus system.

Junctions

Non-parent junctions that can be constructed, up to the length of 8. This specific Junctions implementation uses a Rust enum in order to make pattern matching easier.

MaybeErrorCode

NetworkId

A global identifier of a data structure existing within consensus.

OriginKind

Basically just the XCM (more general) version of ParachainDispatchOrigin.

Outcome

Outcome of an XCM execution.

Response

Response data to a query.

SendError

Error result value when attempting to send an XCM message.

Unsupported

VersionedAsset

A single Asset value, together with its version code.

VersionedAssetId

A single version’s AssetId value, together with its version code.

VersionedAssets

A single MultiAssets value, together with its version code.

VersionedInteriorLocation

A single InteriorLocation value, together with its version code.

VersionedLocation

A single Location value, together with its version code.

VersionedResponse

A single version’s Response value, together with its version code.

VersionedXcm

A single XCM message, together with its version code.

WeightLimit

An optional weight limit.

WestendOriginCaller

WildAsset

A wildcard representing a set of assets.

WildFungibility

Classification of whether an asset is fungible or not.

XcmError

Error codes used in XCM. The first errors codes have explicit indices and are part of the XCM format. Those trailing are merely part of the XCM implementation; there is no expectation that they will retain the same index over time.

Constants

ALICE

ASSET_HUB_WESTEND_ED

COLLECTIVES_WESTEND_ED

MIN_XCM_VERSION

The minimal supported XCM version

WESTEND_ED

XCM_VERSION

Traits

AssetHubWestendPallet

Chain

CollectivesWestendPallet

ExecuteXcm

Type of XCM message executor.

GetVersion

Check and return the Version that should be used for the Xcm datum for the destination Location, which will interpret it.

IdentifyVersion

Used to get the version out of a versioned type.

IntoVersion

Attempt to convert self into a particular version of itself.

Parachain

PreparedMessage

Reanchorable

Relay

SendXcm

Utility for sending an XCM message to a given location.

TestExt

WestendPallet

WrapVersion

Convert an Xcm datum into a VersionedXcm, based on a destination Location which will interpret it.

XcmWeightInfo

Functions

send_xcm

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as mutable references into T::send_xcm.

validate_send

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as mutable references into T::send_xcm.

Type Aliases

AlwaysLatest

WrapVersion implementation which attempts to always convert the XCM to the latest version before wrapping it.

AlwaysLts

WrapVersion implementation which attempts to always convert the XCM to the most recent Long- Term-Support version before wrapping it.

AssetHubLocationToAccountId

Type for specifying how a Location can be converted into an AccountId. This is used when determining ownership of accounts for asset transacting and when attempting to use XCM Transact in order to determine the dispatch Origin.

AssetHubWestend

CollectivesWestend

InteriorLocation

A relative location which is constrained to be an interior location of the context.

QueryId

An identifier for a query.

SendResult

Result value when attempting to send an XCM message.

Westend

XcmHash

A hash type for identifying messages.

XcmResult

XcmVersion

A version of XCM.