Module collectives_westend_integration_tests::collectives_fellowship

The Westend Technical Fellowship.

Modules

pallet_fellowship_origins

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

ranks

The Fellowship members’ ranks.

Structs

Architects

Burn

EnsureCanPromoteTo

EnsureCanRetainAt

EnsureFellowship

Fellows

FellowshipTreasuryInteriorLocation

FellowshipTreasuryPalletId

HundredPercent

Interior

Masters

MaxBalance

Members

SelfParaId

ToVoice

A TryMorph implementation which is designed to convert an aggregate RuntimeOrigin value into the Fellowship voice it represents if it is a Fellowship pallet origin an appropriate variant. See also Origin::as_voice.

Type Aliases

FellowshipCollectiveInstance

FellowshipCoreInstance

FellowshipReferendaInstance

FellowshipSalaryInstance

FellowshipSalaryPaymaster

PayOverXcm setup to pay the Fellowship salary on the AssetHub in USDT.

FellowshipTreasuryInstance

FellowshipTreasuryPaymaster

PayOverXcm setup to pay the Fellowship Treasury.