Function collectives_westend_integration_tests::validate_send

pub fn validate_send<T>(
    dest: Location,
    msg: Xcm<()>
) -> Result<(<T as SendXcm>::Ticket, Assets), SendError>where
    T: SendXcm,

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as as mutable references into T::send_xcm.