Struct collectives_westend_integration_tests::Assets

pub struct Assets(/* private fields */);

A Vec of Assets.

There are a number of invariants which the construction and mutation functions must ensure are maintained:

It may contain no items of duplicate asset class;

All items must be ordered;

The number of items should grow no larger than MAX_ITEMS_IN_ASSETS.

Implementations

impl Assets

pub fn new() -> Assets

A new (empty) value.

pub fn from_sorted_and_deduplicated(r: Vec<Asset>) -> Result<Assets, ()>

Create a new instance of Assets from a Vec<Asset> whose contents are sorted and which contain no duplicates.

Returns Ok if the operation succeeds and Err if r is out of order or had duplicates. If you canโt guarantee that r is sorted and deduplicated, then use From::<Vec<Asset>>::from which is infallible.

pub fn from_sorted_and_deduplicated_skip_checks(r: Vec<Asset>) -> Assets

Create a new instance of Assets from a Vec<Asset> whose contents are sorted and which contain no duplicates.

In release mode, this skips any checks to ensure that r is correct, making it a negligible-cost operation. Generally though you should avoid using it unless you have a strict proof that r is valid.

In test mode, this checks anyway and panics on fail.

pub fn push(&mut self, a: Asset)

Add some asset onto the list, saturating. This is quite a laborious operation since it maintains the ordering.

pub fn is_none(&self) -> bool

Returns true if this definitely represents no asset.

pub fn contains(&self, inner: &Asset) -> bool

Returns true if self is a super-set of the given inner asset.

pub fn drain(self) -> Vec<Asset>

๐Deprecated: Use into_inner() instead

Consume self and return the inner vec.

pub fn into_inner(self) -> Vec<Asset>

Consume self and return the inner vec.

pub fn inner(&self) -> &Vec<Asset>

Return a reference to the inner vec.

pub fn len(&self) -> usize

Return the number of distinct asset instances contained.

pub fn prepend_with(&mut self, prefix: &Location) -> Result<(), ()>

Prepend a Location to any concrete asset items, giving it a new root location.

pub fn get(&self, index: usize) -> Option<&Asset>

Return a reference to an item at a specific index or None if it doesnโt exist.