Type Alias collectives_westend_integration_tests::AssetHubLocationToAccountId

pub type AssetHubLocationToAccountId = (ParentIsPreset<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>, SiblingParachainConvertsVia<Sibling, <<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>, AccountId32Aliases<RelayNetwork, <<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>, HashedDescription<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId, DescribeFamily<DescribePalletTerminal>>, GlobalConsensusParachainConvertsFor<UniversalLocation, <<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>);

Type for specifying how a Location can be converted into an AccountId. This is used when determining ownership of accounts for asset transacting and when attempting to use XCM Transact in order to determine the dispatch Origin.